backend/app/requirements.txt
fastapi~=0.110.1
SQLAlchemy~=2.0.29
psycopg2-binary~=2.9.1
requests~=2.31.0
python-jose~=3.3.0
uvicorn~=0.29.0
python-multipart~=0.0.9
fastanpr~=0.1.13
httptools==0.6.1
scikit-learn==1.4.2
tensorflow==2.16.1
lapx~=0.5.9